ExchangeSystem/cosmos-app/requirements.txt
Django==4.1.7
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.5
ecdsa==0.17.0
frozenlist==1.3.3
furl==2.1.3
h11==0.14.0
h2==4.1.0
hdwallets==0.1.2
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
multidict==6.0.4
nest-asyncio==1.5.6
openapi-codec==1.3.2
packaging==23.0
psycopg2-binary==2.9.5
pycryptodome==3.17
python-dateutil==2.8.2
rfc3986==1.5.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.18.3
sniffio==1.3.0
uritemplate==4.1.1
wrapt==1.14.1
yarl==1.8.2
async-timeout==4.0.2
django-redis==5.2.0
redis==4.5.1
asn1crypto==1.5.1
attrs==22.2.0
blockfrost-python==0.5.2
cbor2==5.4.6
certifi==2022.12.7
certvalidator==0.11.1
cffi==1.15.1
charset-normalizer==3.0.1
cose==0.9.dev8
cryptography==39.0.1
ECPy==1.2.5
idna==3.4
oscrypto==1.3.0
pprintpp==0.4.0
pycardano==0.7.3
pycparser==2.21
PyNaCl==1.5.0
requests==2.28.2
six==1.16.0
typeguard==2.13.3
urllib3==1.26.14
websocket-client==1.5.1
mnemonic==0.20
asgiref==3.6.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3